rename typescript entity files to non-plural to match the class names inside.
